# Optimise BiasedRandomWalk