- Weighted:
  - This PR: 9.03s
  - 991db681: 34.4s
  - f2a96aab: 

Unweighted:
  - This PR:
  - 991db681: 6.58s
  - f2a96aab: 4.37s

See: #1441